refactored test_sphere to test directory.
